Split build support into crate
