+++ crates/jcode-build-support/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "jcode-build-support"
+version = "0.1.0"
+edition = "2024"
+publish = false
+
+[dependencies]
+anyhow = "1"
+chrono = { version = "0.4", features = ["serde"] }
+jcode-core = { path = "../jcode-core" }
+jcode-selfdev-types = { path = "../jcode-selfdev-types" }
+jcode-storage = { path = "../jcode-storage" }
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
+tempfile = "3"
+
+[dev-dependencies]
+tempfile = "3"
